moved fetch-hook to parent component and updated imports and types
